terminado venta de acciones , arreglados bugs de tabla que generaba row vacia y redondeo de prices , cambie algunos eventos no me acuerdo exactamente cuales key released por onclick era uno para arreglar algunas boludeces. Y estube testeando desde Inicio.. no se si tenes que cambiar algo en netbeans

+++ src/main/java/controlers/AccionsController.java
@@ -32,7 +32,7 @@ public class AccionsController {
     }
 
     // AGREGAR ACCION.
-    public void comprarAccion(String codClient, String company, int price, int quantity){
+    public void comprarAccion(String codClient, String company, float price, int quantity){
 
         // busca las acciones del cliente
         List<Accion> list = this.traerAccionesCliente(codClient);
